Added proper functionality to simple plots.

+++ CerebellumProjectModelling.py
@@ -12,6 +12,11 @@ from matplotlib.lines import Line2D
 Saves simple and logged plots to separate files. Will by default open windows with each figure,
 comment out 'plt.show()' at end of file to save only."""
 
+# TODO: Make it so you can choose whether you want to save (with user input)
+#  Make function to let people clear the save folder when called.
+#  Requirements.txt
+#  Fix input bug - (needs to return back to the initial user input). IN ELSE BLOCK.
+
 try:
     data = pd.read_csv('all_species_values.csv', na_values='', usecols=range(7))
     data = data.dropna(how='all', axis='columns').drop(columns='Source')
@@ -100,21 +105,35 @@ def plot_variables(xy=var_combinations, logged=False):
 
             fig1.tight_layout(pad=0.4, w_pad=0.5, h_pad=1.0)
 
-        save_folder = os.path.join(os.getcwd(), r'Saved Simple Plots')
+        save_check = input('Do you wish to save this plot? (Y/N) ')
 
-        if not os.path.exists(save_folder):
-            os.makedirs(save_folder)
+        if save_check.lower() == 'y':
+            save_folder = os.path.join(os.getcwd(), r'Saved Simple Plots')
+            if not os.path.exists(save_folder):
+                os.makedirs(save_folder)
 
-        png_id = 0
+            png_id = 0
 
-        if xy is var_combinations:
-            while os.path.exists(f'Saved Simple Plots/{"Simple Default Plots - "}{png_id:d}.png'):
-                png_id += 1
-            plt.savefig(f'Saved Simple Plots/{"Simple Default Plots - "}{png_id:d}.png')
+            if xy is var_combinations:
+                while os.path.exists(f'Saved Simple Plots/{"Simple Default Plots - "}{png_id:d}.png'):
+                    png_id += 1
+                plt.savefig(f'Saved Simple Plots/{"Simple Default Plots - "}{png_id:d}.png')
+
+            else:
+                while os.path.exists(f'Saved Simple Plots/{len(xy)} {"Simple Plot(s) - "}{png_id:d}.png'):
+                    png_id += 1
+                plt.savefig(f'Saved Simple Plots/{len(xy)} {"Simple Plot(s) - "}{png_id:d}.png')
+
+            print('Plots Saved!')
+            sys.exit()
+
+        elif save_check.lower() == 'n':
+            print('Figures not saved. '
+                  'Call plt.show() to instead output figures to a new window.')
+            sys.exit()
+
         else:
-            while os.path.exists(f'Saved Simple Plots/{len(xy)} {"Simple Plot(s) - "}{png_id:d}.png'):
-                png_id += 1
-            plt.savefig(f'Saved Simple Plots/{len(xy)} {"Simple Plot(s) - "}{png_id:d}.png')
+            print('Invalid Input. Enter Y or N: ')
 
     elif logged:
         fig2, axs2 = plt.subplots(1, len(xy), figsize=(fig_width, 4), squeeze=False)
@@ -167,7 +186,7 @@ def plot_variables(xy=var_combinations, logged=False):
                     png_id += 1
                 plt.savefig(f'Saved Log Plots/{len(xy)} {"Logged Plot(s) - "}{png_id:d}.png')
 
-            print('Saved!')
+            print('Plots Saved!')
             sys.exit()
 
         elif save_check.lower() == 'n':
@@ -179,9 +198,9 @@ def plot_variables(xy=var_combinations, logged=False):
             print('Invalid Input. Enter Y or N: ')
 
 
-# plot_variables()
-plot_variables(logged=True)
-
-plot_variables((('Cerebrum Volume', 'Cerebellum Volume'),), logged=True)
+plot_variables()
+# plot_variables(logged=True)
+#
+# plot_variables((('Cerebrum Volume', 'Cerebellum Volume'),), logged=True)
 
 # plt.show()
